requirements.txt
inference
supervision
python-dotenv
opencv-python
flask
roboflow
